374a368 Fix gdcm warnings emitted by the Microsoft compiler.

+++ Utilities/gdcm/Source/MediaStorageAndFileFormat/gdcmJPEGCodec.cxx
@@ -37,7 +37,7 @@ JPEGCodec::~JPEGCodec()
 
 void JPEGCodec::SetQuality(double q)
 {
-  Quality = q;
+  Quality = static_cast<int>(q);
 }
 
 double JPEGCodec::GetQuality() const
